move token exchange from plugin to openai client
